+++ client/src/lib/hooks/graph.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { GraphContext } from './graph.tsx'
+
+export const useGraph = () => useContext(GraphContext)
