+++ client/src/lib/hooks/graph.tsx
@@ -1,4 +1,4 @@
-import { createContext, FC, ReactNode, useContext, useEffect, useMemo, useRef, useState } from 'react'
+import { createContext, FC, ReactNode, useEffect, useMemo, useRef, useState } from 'react'
 import styled from 'styled-components'
 import { Edge, Node } from '../api/api_pb.ts'
 import { GraphBuilder } from '../graph/graphBuilder.ts'
@@ -60,20 +60,6 @@ export const GraphWrapper = styled.div`
   left: 0;
   z-index: 9;
 `
-
-export const useGraph = () => useContext(GraphContext)
-
-export const useGraphBuilder = (nodes: Node[], edges: Edge[], onShiftClick?: (gb: GraphBuilder, nodeId: string) => void): GraphBuilder | undefined => {
-  const [graph, setGraph] = useState<GraphBuilder>()
-  useEffect(() => {
-    const gb = new GraphBuilder(window.innerWidth, window.innerHeight, onShiftClick)
-    nodes.forEach(node => gb.addNode(node))
-    edges.forEach(edge => gb.addEdge(edge))
-    setGraph(gb)
-    return () => setGraph(undefined)
-  }, [nodes, edges, onShiftClick])
-  return graph
-}
 
 export const Graph: FC<{ gb?: GraphBuilder }> = ({ gb }) => {
   const graphWrapperRef = useRef<HTMLDivElement>(null)
